backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
Deprecated==1.2.15
exceptiongroup==1.2.2
fastapi==0.115.6
flatbuffers==24.12.23
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lazy_loader==0.4
limits==4.0.1
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt-toolkit==3.0.48
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
PyMatting==1.1.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
redis==5.2.1
referencing==0.36.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tifffile==2024.8.30
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
